tennis_booking/requirements.txt
selenium==4.15.2
selenium>=4.15.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
pytest==7.4.3
boto3>=1.26.0
